llm_service/requirements.txt
pylint==3.2.5
black==24.4.2
pytest==8.2.2
pytest-cov==5.0.0
pytest-asyncio==0.23.7
ipython==8.26.0
pillow==10.4.0
python-multipart==0.0.9
python-dotenv==1.0.1
langchain==0.2.6
langchain-anthropic==0.1.19
langchain-openai==0.1.14
pika==1.3.2
langchain-cli==0.0.25
langserve[all]